requirements-dev.txt
-r src/requirements.txt

# Cookiecutter Support
cruft
pip-tools

# Testing Tools
pytest
pytest-django
pytest-playwright

# Linters
ruff
